update thubnail components
